src/research-san-pham-moi/requirements.txt
python-docx>=1.1.0
python-pptx>=1.0.0
PyYAML>=6.0
